Atualizar produto com imagens

+++ LojaVirtual/LojaVirtual/Controllers/Libraries/Arquivo/GerenciadorArquivo.cs
@@ -42,22 +42,39 @@ namespace LojaVirtual.Controllers.Libraries.Arquivo
             var ImagensDefinitivo = new List<Imagem>();
             foreach (var CaminhoTemp in caminhosTemp)
             {
-                var NomeArquivo = Path.GetFileName(CaminhoTemp);                
-                var CaminoAbsolutoTemp = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads/temp", NomeArquivo);
-                var CaminhoAbsolutoDefinitivo = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads", idProduto.ToString(), NomeArquivo);
+                var NomeArquivo = Path.GetFileName(CaminhoTemp);
 
-                if (File.Exists(CaminoAbsolutoTemp))
+                var CaminhoDefinitivo = Path.Combine("/uploads", idProduto.ToString(), NomeArquivo).Replace(@"\", "/");
+
+                var caminhoRetorno = new Imagem { Caminho = Path.Combine("/uploads", idProduto.ToString(), NomeArquivo).Replace(@"\", "/"), Id = idProduto };
+                if (CaminhoDefinitivo != CaminhoTemp)
                 {
-                    File.Copy(CaminoAbsolutoTemp, CaminhoAbsolutoDefinitivo);
+                    var CaminoAbsolutoTemp = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads/temp", NomeArquivo);
+                    var CaminhoAbsolutoDefinitivo = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/uploads", idProduto.ToString(), NomeArquivo);
 
-                    if (File.Exists(CaminhoAbsolutoDefinitivo))
+                    if (File.Exists(CaminoAbsolutoTemp))
                     {
-                        File.Delete(CaminoAbsolutoTemp);
+                        //Deleta arquivo na pasta destino
+                        if (File.Exists(CaminhoAbsolutoDefinitivo))
+                        {
+                            File.Delete(CaminhoAbsolutoDefinitivo);
+                        }
 
-                        var caminhoRetorno = new Imagem{ Caminho = Path.Combine("/uploads", idProduto.ToString(), NomeArquivo).Replace(@"\", "/"), Id = idProduto};
-                        ImagensDefinitivo.Add(caminhoRetorno);
+                        //Copira arquivo da pasta temp para a definitiva
+                        File.Copy(CaminoAbsolutoTemp, CaminhoAbsolutoDefinitivo);
+
+                        //Deleta arquivo da pasta temp
+                        if (File.Exists(CaminhoAbsolutoDefinitivo))
+                        {
+                            File.Delete(CaminoAbsolutoTemp);
+                        }
                     }
+                    //else
+                    //{
+                    //    return null;
+                    //}
                 }
+                ImagensDefinitivo.Add(caminhoRetorno);
             }
 
             return ImagensDefinitivo;
